requirements/base.txt
# Base requirements — shared across all environments
Django>=5.2,<5.3
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.4,<6
psycopg[binary]>=3.2,<4
django-redis>=5.4,<6
django-storages[s3]>=1.14,<2
Pillow>=11,<12
celery>=5.4,<6
django-celery-beat>=2.7,<3
redis>=5.2,<6
django-cors-headers>=4.6,<5
bleach>=6.2,<7
gunicorn>=23,<24
sentry-sdk[django]>=2,<3
python-decouple>=3.8,<4
dj-database-url>=2.3,<3
drf-spectacular>=0.28,<1
python-json-logger>=3,<5
africastalking>=2.0,<3
channels>=4.2,<5
daphne>=4.1,<5
channels-redis>=4.2,<5
